Implementing the Session:all_dbs method
This method makes an http request to the _all_dbs http uri of the
couchdb uri and decodes the returned json value. This commit makes
lua-couchdb depend on lua-json and lua-socket2.

Another small change was that now we return the _M var in the end of
the module. And now, Session object is not local anymor, that was
causing an error!