Starting the implementation of the Document data type
To PUT (or POST) a document, we need to send something when doing
a request, so the _do_request() function was changed to receive a new
param, the content.

This new param receives *LUA objects*. They will be converted to json
and then sent to couchdb.

When content is available, a custom header (Content-Type) is also set.